Finally the cubes are getting rendered. DrawGraph working.

+++ Source/Core/ECS/Components/Mesh.h
@@ -31,7 +31,7 @@ public:
     std::vector<glm::vec2*> uv8;
 
     std::vector<glm::vec3*> positions;
-    std::vector<glm::uvec4*> colors;
+    std::vector<glm::vec4*> colors;
     std::vector<glm::vec3*> normals;
     std::vector<glm::vec3*> tangents;
     std::vector<uint32_t*> indicies;
